Fixed system.in uses buffering for tty.
Added the tty method that returns whether it is a tty to io.File.
Added the buffer size argument to the constructor of io.Reader.
Changed to specify the buffer size of system.in according to whether its stream is a tty.
Changed native_{in,out,error} to raw_{in,out,error} respectively.